Take 2
Treat ns/ as sym_lit, but unlike previous
attempt, maintain parsing of symbols further into
components, and thus retaining fields.

A key part of getting this to work was to change
the lexical precedence of SYMBOL_NAMESPACED_NAME
to be higher than that of SYMBOL so that there
was a preference for parsing certain symbols as
a single unit (_sym_qualified), instead of as
two units (_sym_almost + _sym_unqualified).